EPMLABSBRN-191 add level in loading file

+++ src/main/kotlin/com/epam/brn/service/parsers/csv/dto/ExerciseCsv.kt
@@ -12,7 +12,7 @@ data class ExerciseCsv(
     val seriesId: Long,
 
     @JsonProperty("level")
-    val level: Short,
+    val level: Int,
 
     @JsonProperty("name")
     val name: String,
